move validator to handlers package
