+++ handlers/validator.go
@@ -1,4 +1,4 @@
-package models
+package handlers
 
 import (
 	"errors"
